Move platform-neutral part of TimeZone to shared CoreLib partition (#15926)
Prep work for moving the platform specific parts
